requirements.txt
Telethon==1.38.1
python-dotenv==1.0.0
pytest==7.4.2
pandas==2.2.3